requirements.txt
pyspark
confluent-kafka
pyarrow
pyyaml
